requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2020.6.20
click==8.1.3
Deprecated==1.2.13
Django==4.0.4
django-rq==2.5.1
djongo==1.3.6
GitPython==3.1.27
packaging==21.3
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.1
rq==1.10.1
sqlparse==0.2.4
wrapt==1.14.1
